Worked on redirecrion
