--- lib/screens/my_account_page.dart
@@ -1,20 +1,0 @@
-import 'package:flutter/material.dart';
-
-class MyAccountPage extends StatefulWidget {
-  const MyAccountPage({Key? key}) : super(key: key);
-
-  @override
-  State<MyAccountPage> createState() => _MyAccountPageState();
-}
-
-class _MyAccountPageState extends State<MyAccountPage> {
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      child: const Align(
-        alignment: Alignment.center,
-        child: Text("My Account"),
-      ),
-    );
-  }
-}
