--- lib/screens/premium_page.dart
@@ -1,18 +1,0 @@
-import 'dart:convert';
-
-import 'package:flutter/material.dart';
-
-class PremiumPage extends StatefulWidget {
-  @override
-  State<PremiumPage> createState() => _PremiumPageState();
-}
-
-class _PremiumPageState extends State<PremiumPage> {
-
-  @override
-  Widget build(BuildContext context) {
-    return Text("Premium");
-
-}
-
-}
