+++ lib/screens/setting_screen.dart
@@ -5,40 +5,33 @@ import 'package:getwidget/types/gf_toggle_type.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:newsdx/app_constants/string_constant.dart';
 import 'package:newsdx/preference/user_preference.dart';
-import 'package:provider/provider.dart';
 import 'package:theme_provider/theme_provider.dart';
 
-class MoreScreen extends StatefulWidget {
-  const MoreScreen({Key? key}) : super(key: key);
+class MyAccountPage extends StatefulWidget {
+  const MyAccountPage({Key? key}) : super(key: key);
 
   @override
-  State<MoreScreen> createState() => _MoreScreenState();
+  State<MyAccountPage> createState() => _MyAccountPageState();
 }
 
-class _MoreScreenState extends State<MoreScreen> {
+class _MyAccountPageState extends State<MyAccountPage> {
   @override
   Widget build(BuildContext context) {
-    return  Scaffold(
+    return Scaffold(
       backgroundColor: Colors.white,
       appBar: AppBar(
         backgroundColor: Colors.white,
         centerTitle: true,
         elevation: 0.0,
-        leading: Transform.scale(
-            scale: 1.2,
-            child: IconButton(
-                icon: SvgPicture.asset("assets/back.svg"), onPressed: () {
-                  Navigator.of(context).pop();
-            }),),
         actions: [
           Padding(
-            padding: const EdgeInsets.only(right: 16),
-            child: CircleAvatar(
-              radius: 16.0,
-              child: ClipRect(
-                child: Prefs.getProfilePre() ? Image.network('https://picsum.photos/250?image=9') : SvgPicture.asset("assets/profile_placeholder.svg"),
-              ),
-            )
+              padding: const EdgeInsets.only(right: 16),
+              child: CircleAvatar(
+                radius: 16.0,
+                child: ClipRect(
+                  child: Prefs.getProfilePre() ? Image.network('https://picsum.photos/250?image=9') : SvgPicture.asset("assets/profile_placeholder.svg"),
+                ),
+              )
           )
         ],
       ),
@@ -51,9 +44,9 @@ class _MoreScreenState extends State<MoreScreen> {
                 children: [
                   SvgPicture.asset("assets/more.svg", height: 20, width: 20,),
                   const Text(" More", style: TextStyle(
-                    fontSize: 38,
-                    fontWeight: FontWeight.w800,
-                    color: Colors.black
+                      fontSize: 38,
+                      fontWeight: FontWeight.w800,
+                      color: Colors.black
                   ),),
                 ],
               ),
@@ -88,26 +81,26 @@ class _MoreScreenState extends State<MoreScreen> {
             ),),
             Padding(padding: const EdgeInsets.only(left: 16, right: 16, top: 20), child: Column(
               children: [
-               GestureDetector(
-                 child:  Row(
-                   mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                   children: [
-                     const Text("Dark mode", style: TextStyle(
-                       color: Colors.grey,
-                       fontSize: 16,
-                     ),),
-                     GFToggle(
-                       onChanged: (val){},
-                       value: true,
-                       enabledTrackColor: Colors.blue,
-                       type: GFToggleType.ios,
-                     )
-                   ],
-                 ),
-                 onTap: (){
-                   ThemeProvider.controllerOf(context).nextTheme();
-                 },
-               ),
+                GestureDetector(
+                  child:  Row(
+                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                    children: [
+                      const Text("Dark mode", style: TextStyle(
+                        color: Colors.grey,
+                        fontSize: 16,
+                      ),),
+                      GFToggle(
+                        onChanged: (val){},
+                        value: true,
+                        enabledTrackColor: Colors.blue,
+                        type: GFToggleType.ios,
+                      )
+                    ],
+                  ),
+                  onTap: (){
+                    ThemeProvider.controllerOf(context).nextTheme();
+                  },
+                ),
                 const SizedBox(
                   height: 10,
                 ),
@@ -193,31 +186,32 @@ class _MoreScreenState extends State<MoreScreen> {
                 Row(
                   mainAxisAlignment: MainAxisAlignment.spaceBetween,
                   children: [
-                   Row(
-                     children: [
-                       const Text("Notifications", style: TextStyle(
-                         color: Colors.grey,
-                         fontSize: 16,
-                       ),),
-                       const SizedBox(
-                         width: 4,
-                       ),
-                       SvgPicture.asset("assets/dot.svg"),
-                     ],
-                   ),
-                   Row(
-                     children: [
-                       const Text("200 unread",
-                       style: TextStyle(
-                         color: Colors.grey,
-                         fontSize: 14
-                       ),),
-                       const SizedBox(
-                         width: 5,
-                       ),
-                       SvgPicture.asset("assets/forward.svg"),
-                     ],
-                   ),
+                    Row(
+                      children: const [
+                        Text("Notifications", style: TextStyle(
+                          color: Colors.grey,
+                          fontSize: 16,
+                        ),),
+                        SizedBox(
+                          width: 4,
+                        ),
+                        // SvgPicture.asset("assets/dot.svg"),
+                      ],
+                    ),
+                    Row(
+                      children: [
+                        // 200 unread can be added notification
+                        const Text("",
+                          style: TextStyle(
+                              color: Colors.grey,
+                              fontSize: 14
+                          ),),
+                        const SizedBox(
+                          width: 5,
+                        ),
+                        SvgPicture.asset("assets/forward.svg"),
+                      ],
+                    ),
                   ],
                 ),
                 const SizedBox(
@@ -237,20 +231,21 @@ class _MoreScreenState extends State<MoreScreen> {
                   mainAxisAlignment: MainAxisAlignment.spaceBetween,
                   children: [
                     Row(
-                      children: [
-                        const Text("Bookmark", style: TextStyle(
+                      children: const [
+                        Text("Bookmark", style: TextStyle(
                           color: Colors.grey,
                           fontSize: 16,
                         ),),
-                        const SizedBox(
+                        SizedBox(
                           width: 4,
                         ),
-                        SvgPicture.asset("assets/dot.svg"),
+                        // SvgPicture.asset("assets/dot.svg"),
                       ],
                     ),
                     Row(
                       children: [
-                        const Text("200 unread",
+                        // 200 unread can be added for bookmark
+                        const Text("",
                           style: TextStyle(
                               color: Colors.grey,
                               fontSize: 14
@@ -276,7 +271,7 @@ class _MoreScreenState extends State<MoreScreen> {
             ),),
             Column(
               children: [
-                Padding(padding: const EdgeInsets.only(top: 150),
+                Padding(padding: const EdgeInsets.only(top: 100),
                     child: SizedBox(
                       width: 75,
                       height: 20 ,
